renamed ImageEditor and added required methods (crop, resize, greyscale)
