--- app/src/main/java/de/hsaugsburg/teamulster/sohappy/analyzer/ImageEditor.kt
@@ -1,14 +1,0 @@
-package de.hsaugsburg.teamulster.sohappy.analyzer
-
-import android.graphics.Bitmap
-import android.graphics.Rect
-
-@Suppress("UtilityClassWithPublicConstructor")
-class ImageEditor {
-    companion object {
-        @Suppress("FunctionOnlyReturningConstant")
-        fun crop (img: Bitmap, frame: Rect): Bitmap? {
-            return null
-        }
-    }
-}
